Update and rename Notepad++.py to notepad++.py
